// src/app/shared/model/LTKM/LtkmCustGrpObj.Model.ts
export class LtkmCustGrpObj {
    LtkmCustGrpId: number;
    LtkmCustId: number;
    CustNo: string;
    CustName: string;
    MrCustRelationshipCode: string;
    RelationshipName: string;
    CustGrpNotes: string;
    IsReversible: boolean;


    constructor() {
        this.LtkmCustGrpId = 0;
    }
}